requirements.txt
kafka-python
Flask 
Flask-SocketIO 
kafka-python
flask-cors